# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Motorcycle ERP & POS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /grn:
    get:
      operationId: listGRNs
      tags: [grn]
      summary: List goods received notes
      parameters:
        - in: query
          name: purchaseOrderId
          schema:
            type: integer
      responses:
        "200":
          description: List of GRNs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GRN"
    post:
      operationId: createGRN
      tags: [grn]
      summary: Create a GRN and update stock levels
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGRNBody"
      responses:
        "201":
          description: GRN created, stock updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GRN"
  /grn/{id}:
    get:
      operationId: getGRN
      tags: [grn]
      summary: Get GRN by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: GRN with lines
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GRNWithLines"

  # ─── WORK ORDERS ───────────────────────────────────────────────────────────
...
  /inspections:
    get:
      operationId: listInspections
      tags: [inspections]
      summary: List pre-owned inspection reports
      parameters:
        - in: query
          name: motorcycleId
          schema:
            type: integer
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of inspections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Inspection"
    post:
      operationId: createInspection
      tags: [inspections]
      summary: Create an inspection report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInspectionBody"
      responses:
        "201":
          description: Inspection created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inspection"
  /inspections/{id}:
    get:
      operationId: getInspection
      tags: [inspections]
      summary: Get inspection by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inspection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inspection"
    put:
      operationId: updateInspection
      tags: [inspections]
      summary: Update inspection
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInspectionBody"
      responses:
        "200":
          description: Updated inspection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inspection"

  # ─── AUDIT LOG ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    GRN:
      type: object
      properties:
        id:
          type: integer
        grnNumber:
          type: string
        purchaseOrderId:
          type: integer
        poNumber:
          type: string
        vendorId:
          type: integer
        vendorName:
          type: string
        receivedAt:
          type: string
          format: date-time
        receivedBy:
          type: integer
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - grnNumber
        - purchaseOrderId
        - receivedAt
        - createdAt
    GRNLine:
      type: object
      properties:
        id:
          type: integer
        grnId:
          type: integer
        partId:
          type: integer
        partSku:
          type: string
        partName:
          type: string
        quantityReceived:
          type: integer
        binId:
          type: integer
        binLabel:
          type: string
      required:
        - id
        - grnId
        - partId
        - quantityReceived
    GRNWithLines:
      allOf:
        - $ref: "#/components/schemas/GRN"
        - type: object
          properties:
            lines:
              type: array
              items:
                $ref: "#/components/schemas/GRNLine"
    CreateGRNBody:
      type: object
      properties:
        purchaseOrderId:
          type: integer
        notes:
          type: string
        lines:
          type: array
          items:
            $ref: "#/components/schemas/CreateGRNLineBody"
      required:
        - purchaseOrderId
        - lines
    CreateGRNLineBody:
      type: object
      properties:
        partId:
          type: integer
        quantityReceived:
          type: integer
        binId:
          type: integer
      required:
        - partId
        - quantityReceived

    # ─── WORK ORDER ──────────────────────────────────────────────────────────
...
    Inspection:
      type: object
      properties:
        id:
          type: integer
        motorcycleId:
          type: integer
        motorcycleDesc:
          type: string
        inspectorId:
          type: integer
        inspectorName:
          type: string
        overallGrade:
          type: string
          enum: [excellent, good, fair, poor]
        engineCondition:
          type: string
        bodyCondition:
          type: string
        electricalCondition:
          type: string
        tiresCondition:
          type: string
        brakeCondition:
          type: string
        notes:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        isCertified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - motorcycleId
        - overallGrade
        - isCertified
        - createdAt
    CreateInspectionBody:
      type: object
      properties:
        motorcycleId:
          type: integer
        overallGrade:
          type: string
          enum: [excellent, good, fair, poor]
        engineCondition:
          type: string
        bodyCondition:
          type: string
        electricalCondition:
          type: string
        tiresCondition:
          type: string
        brakeCondition:
          type: string
        notes:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        isCertified:
          type: boolean
      required:
        - motorcycleId
        - overallGrade
    UpdateInspectionBody:
      type: object
      properties:
        overallGrade:
          type: string
          enum: [excellent, good, fair, poor]
        engineCondition:
          type: string
        bodyCondition:
          type: string
        electricalCondition:
          type: string
        tiresCondition:
          type: string
        brakeCondition:
          type: string
        notes:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        isCertified:
          type: boolean

    # ─── AUDIT LOG ───────────────────────────────────────────────────────────
...